fix: biginteger tg_id

+++ db/models/profiles.py
@@ -1,6 +1,6 @@
 from contextvars import ContextVar
 
-from sqlalchemy import Column, Integer
+from sqlalchemy import Column, BigInteger
 
 from db.models.base import ModelBase
 
@@ -8,7 +8,7 @@ from db.models.base import ModelBase
 class Profiles(ModelBase):
     __tablename__ = "profiles"
 
-    tg_id = Column(Integer, unique=True, index=True, nullable=False)
+    tg_id = Column(BigInteger, unique=True, index=True, nullable=False)
 
 
 profile_session_var: ContextVar[Profiles] = ContextVar("profile_session_var")
